% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoterUtils.R
\name{connect_remote}
\alias{connect_remote}
\title{A \code{remoter} Wrapper}
\usage{
connect_remote(
  credentials = NULL,
  prompt = "REMOTE_SESSION::",
  port = NULL,
  action = client,
  session = make.names(tolower(Sys.getenv("COMPUTERNAME"))),
  ...
)
}
\arguments{
\item{credentials}{One of the following:
\itemize{
\item{\code{string}: The path to the R workspace image containing the cipher and key needed to connect to the remote resource}
\item{\code{TRUE}: Choose the R workspace image interactively via \code{\link[tcltk]{tk_choose.files}}}
\item{'ENV': See 'Details'}
\item{\code{NULL|FALSE}: Use when no credentials are needed (connections will \emph{not} be secure)}
}}

\item{prompt}{(string) The prompt to use during the remote session}

\item{port}{(integer) The port to use if not provided in the cipher (useful for specific overrides or ad-hoc sessions)}

\item{action}{(string, symbol) The function to use to connect to the remote session (e.g., \code{client}, \code{batch}, etc.)}

\item{session}{(string) The name of the target session when \code{credentials} is \emph{'ENV'}}

\item{...}{Not used}
}
\description{
\code{connect_remote} is a wrapper for \code{remoter::\link[remoter]{client}} making use of predefined encrypted authentication objects
}
\details{
When \code{credentials} is passed \emph{'ENV'}, a Windows environment variable named 'remoter_<cipher prefix>' must already be set.
The content of the variable should be a JSON string created with \code{\link{make_cipher_env}}.  Expected keys are as follows:
\itemize{
\item{addr: string (IPV4 format)}
\item{port: numeric,string}
\item{session: string}
\item{cipher: string (hexidecimal)}
\item{nonce: string (hexidecimal, must evaluate to length 24L raw vector)}
\item{shared_key: string (hexidecimal)}
}
}
\section{References}{

At a minimum, read the following documentation:
\itemize{
\item{\code{\link[remoter]{client}}}
\item{\code{\link[remoter]{c2s}}}
\item{\code{\link[remoter]{s2c}}}
\item{\code{\link[remoter]{exit}}}
}
}

