% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoterUtils.R
\name{gen_pass}
\alias{gen_pass}
\title{Generate a Password}
\usage{
gen_pass(glyphs = "@$", length = NULL, raw = FALSE, chatty = FALSE)
}
\arguments{
\item{glyphs}{Character-coercibles to use in the creation of the password: this is combined with the output of \code{\link[sodium]{keygen}}}

\item{length}{(int) The length of the password in character format}

\item{raw}{(logical) Should the output be returned as raw?}

\item{chatty}{(logical) Should diagnostic information be provided?}
}
\description{
\code{gen_pass} creates a password consisting of alphanumeric glyphs and symbols
}
\note{
The generated string always begins with a letter before being returned as-is or returned as a raw vector
}
