% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cipher.R
\name{hostname}
\alias{hostname}
\title{Create a Hostname}
\usage{
hostname(x = NULL)
}
\arguments{
\item{x}{(name|string) The name of the host}
}
\value{
A string for use with \code{\link{make_cipher}}
}
\description{
Create a Hostname
}
