% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_batch_file.R
\name{make.batch_file}
\alias{make.batch_file}
\title{Make a \code{remoter} Server Batch File}
\usage{
make.batch_file(
  server_dir,
  auth_root,
  batch.file_name = "start_Rserver.bat",
  source.file_name = "Rserve.R",
  session = make.names(tolower(Sys.getenv("COMPUTERNAME")))
)
}
\arguments{
\item{server_dir}{(string) The server working directory}

\item{auth_root}{(string) The path to the authentication objects}

\item{batch.file_name}{(string) The name of the batch file to be created in \code{server_dir}}

\item{source.file_name}{(string) The name of the R source file to be created in \code{server_dir}}

\item{session}{(string) The session label for the spawned server: must match the cipher object that is ultimately loaded}
}
\value{
\code{source.file_name} and \code{batch.file_name} (which calls \code{source.file_name}) created in \code{server_dir}
}
\description{
\code{make.batch_file} creates the Windows OS batch file and associated R script to launch the \code{remoter} server.  Existing contents in \code{server_dir} are deleted before repopulating.
}
