% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoterUtils.R
\name{make_batch_file}
\alias{make_batch_file}
\title{Make a \code{remoter} Server Batch File}
\usage{
make_batch_file(
  server_dir,
  auth_root,
  batch.file_name = "start_Rserver.bat",
  source.file_name = "Rserve.R",
  sessOpts = make.names(tolower(Sys.getenv("COMPUTERNAME")))
)
}
\arguments{
\item{server_dir}{(string) The server working directory}

\item{auth_root}{(string) The path to the authentication objects}

\item{batch.file_name}{(string) The name of the batch file to be created in \code{server_dir}}

\item{source.file_name}{(string) The name of the R source file to be created in \code{server_dir}}

\item{sessOpts}{(string[]) The session labels expected by the spawned server}
}
\value{
\code{source.file_name} and \code{batch.file_name} (which calls \code{source.file_name}) created in \code{server_dir}
}
\description{
Make a \code{remoter} Server Batch File
}
\note{
Existing contents in \code{server_dir} are deleted before repopulating
}
\references{
\href{https://superuser.com/questions/149951/does-in-batch-file-mean-all-command-line-arguments}
}
