% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoterUtils.R
\name{make_cipher}
\alias{make_cipher}
\title{Create \code{remoter} Authorization Objects}
\usage{
make_cipher(
  file_prefix = make.names(tolower(Sys.getenv("COMPUTERNAME"))),
  host_ip,
  host_port,
  password,
  shared_key,
  export = FALSE
)
}
\arguments{
\item{file_prefix}{(string) The prefix for the output file in the user directory having suffix "_remoter_auth.rdata"}

\item{host_ip}{(string) The IPV4 address for the server (hostnames are not allowed)}

\item{host_port}{(string,integer) The port to use for the server: should be a static, reserved port}

\item{password}{(string) The password for the server}

\item{shared_key}{(string) The key to encrypt and decrypt the generated cipher}

\item{export}{(logical | \code{FALSE}) When \code{TRUE}, the cipher will be saved to the user directory as "~/"\code{file_prefix}_\code{remoter_auth.rdata}"}
}
\value{
The cipher (invisibly) having required attributes \code{nonce}, \code{port}, and \code{addr}
}
\description{
\code{remoter_auth} creates an encrypted object used to securely start and connect to a \code{\link[remoter]{remoter}} session
}
