% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cipher.R
\name{make_cipher_env}
\alias{make_cipher_env}
\title{Make a Cipher OS Environment Variable}
\usage{
make_cipher_env(
  cipher = NULL,
  shared_key = NULL,
  session = make.names(tolower(Sys.getenv("COMPUTERNAME")))
)
}
\arguments{
\item{cipher}{If \code{NULL},the R image holding encryption objects is interactively chosen; otherwise, an object created with \code{\link{make_cipher}}.}

\item{shared_key}{If the shared key was not included in the cipher object, this value is checked and throws an error if empty}

\item{session}{The name of the session for which the environment variable string is created}
}
\value{
a JSON string that can be stored in a Windows OS environment variable formatted as "remoter_\if{html}{\out{<session>}}"
}
\description{
\code{make_cipher_env}() creates a Windows OS environment variable string from the provided cipher. OS environment variables beginning with 'remoter_' and ending with the target session string are searched when calling \code{\link{connect_remote}} with \code{credentials = "ENV"}.
}
