% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoterUtils.R
\name{remoterUtils}
\alias{remoterUtils}
\title{Utility Functions for \code{remoter}}
\description{
Package \code{remoterUtils} provides utility functions facilitating management of \code{remoter} sessions as well as authentication objects (see \code{\link[remoter]{remoter-package}} for \code{remoter} documentation).
}
\section{GitHub Dependencies}{

\enumerate{
\item{remotes::install_github("wrathematics/argon2")}
\item{remotes::install_github("RBigData/remoter")}
}
}

