% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoterUtils.R
\name{search}
\alias{search}
\title{Override \code{search()}}
\usage{
search()
}
\description{
This function overrides the default \code{\link[base]{search}} function to allow for use in \href{https://github.com/RBigData/remoter}{remoter} sessions.
}
\seealso{
Other Overrides: 
\code{\link{searchpaths}()}
}
\concept{Overrides}
