% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoterUtils.R
\name{searchpaths}
\alias{searchpaths}
\title{Override \code{searchpaths()}}
\usage{
searchpaths()
}
\description{
This function overrides the default \code{\link[base]{searchpaths}} function for use in \href{https://github.com/RBigData/remoter}{remoter} sessions.
}
\seealso{
Other Overrides: 
\code{\link{search}()}
}
\concept{Overrides}
