% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cipher.R
\name{make_cipher}
\alias{make_cipher}
\title{Create \code{remoter} Authorization Objects}
\usage{
make_cipher(
  file_prefix = hostname(),
  hostname = hostname(),
  host_ip,
  host_port,
  password,
  shared_key,
  export = FALSE
)
}
\arguments{
\item{file_prefix}{(string) The prefix for the output file in the user directory having suffix "_remoter_auth.rdata"}

\item{hostname}{(string) The label to use for the server}

\item{host_ip}{(string) \emph{(required)} The IPV4 address for the server (hostnames are not allowed).}

\item{host_port}{(string,integer) \emph{(required)} The port to use for the server: should be a static, reserved port}

\item{password}{(string) \emph{(required)} The password for the server}

\item{shared_key}{(string) \emph{(required)} The key to encrypt and decrypt the generated cipher}

\item{export}{(logical | \code{FALSE}) When \code{TRUE}, the cipher will be saved to the user directory as "~/"\code{file_prefix}_\code{remoter_auth.rdata}"}
}
\value{
The cipher (invisibly) having required attributes \code{nonce}, \code{port}, and \code{addr}
}
\description{
\code{make_cipher} creates an encrypted object used to securely start and connect to a \code{remoter} session (see \code{\link[remoter]{remoter-package}} for documentation)
}
