% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoterUtils.R
\name{remoterUtils}
\alias{remoterUtils}
\title{Utility Functions for \code{remoter}}
\description{
\code{remoterUtils} provides utility functions facilitating management of \code{remoter} sessions as well as authentication objects (see \code{\link[remoter]{remoter-package}} for \code{remoter} documentation).  Code is designed for use in a Windows environment.
}
