% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoterUtils.R
\name{server_fun}
\alias{server_fun}
\title{Manage a \code{remoter} Session}
\usage{
server_fun(
  auth_root = path.expand("~"),
  server_dir = "~",
  sessOpts = make.names(tolower(Sys.getenv("COMPUTERNAME"))),
  ...
)
}
\arguments{
\item{server_dir}{The path to the working directory for th spawned server}

\item{sessOpts}{(string[]) Prefixes for the ciphers found in \code{auth_root}: these are the session names that are matched with cipher object names.}

\item{...}{Not used}
}
\description{
\code{server_fun} starts or stops a \code{\link[remoter]{remoter}} session
}
